Address PR review feedback on comparison selection and hooks.
Move the selection dot closer to the card corner and relocate useComparisonSections out of the metrics index file.
